requirements.txt
# confluence_bot dependencies
# Python 3.11+  |  No pandas, no TA-Lib

# Numeric (regime detector ADX, EMA)
numpy>=1.26.0

# Async HTTP
aiohttp>=3.9.0

# WebSockets
websockets>=12.0

# FastAPI metrics server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Config parsing
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Type stubs (dev only)
# mypy>=1.10.0
